third_party/WebKit/LayoutTests/inspector/console/console-proxy-expected.txt

CONSOLE MESSAGE: line 20: [object Proxy]
Tests that console logging dumps proxy properly.

warning-note display: inline-block
info-note display: none
console-proxy.html:20 Proxy console-message-text source-code > console-message-url webkit-html-resource-link > object-value-proxy source-code > console-view-object-properties-section > tree-outline-disclosure tree-outline-disclosure-hide-overflow > tree-outline source-code object-properties-section > parent object-properties-section-root-element > selection fill > object-state-note warning-note > object-state-note info-note > children
window.accessedGet = false
warning-note display: none
info-note display: inline-block
console-proxy.html:20 Proxy__proto__: Object console-message-text source-code > console-message-url webkit-html-resource-link > object-value-proxy source-code > console-view-object-properties-section expanded > tree-outline-disclosure tree-outline-disclosure-hide-overflow > tree-outline source-code object-properties-section > parent object-properties-section-root-element expanded > selection fill > object-state-note warning-note > object-state-note info-note > children expanded > parent > selection fill > name object-properties-section-dimmed > object-properties-section-separator > value object-value-object > children
window.accessedGet = true
